refactor: Put all graphql query documents into documents file.
