+++ tsconfig.json
@@ -14,7 +14,7 @@
     "esModuleInterop": true,
     "exactOptionalPropertyTypes": true,
     "forceConsistentCasingInFileNames": true,
-    "importsNotUsedAsValues": "error",
+    "importsNotUsedAsValues": "preserve",
     "incremental": true,
     "isolatedModules": true,
     "jsx": "preserve",
